+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Unit and E2E tests
+- Server response body is now surfaced on non-success status codes
 
 ### Changed